feat: research on pull-approach
Pull approach - when it is for the consumers to decide, what should be calculated
Pros: easy to find, what work actually must be done and what can be ingonred safely - good for perf at large apps
Cons: hard to actually schedule computations - this is the reason why mobx and similiar are basically using callstack for this, which leads to some tradeoffs

Effector uses push - when every node knows, what should be next
Pros: easier to build effecient calculation
Cons: hard to detect, which work can be ignored safely - bad for large apps

Push-pull difference also known as cold-warm-hot reactions stuff

Big brain idea here is:
make effector nodes to know not just next steps, but previous too
due to already "static" nature of effector it becomes trivial to go up from every effect to find out, which nodes will actually have effects and which are not (and thus can be safely ignored)
this way effector may become "warm" instead of "hot" and looks like changes are not very hard

it is still a wip + just a research

Problems:
Effects are statically defined via operators, so it is ieasy to handle them, but watchers can be added to any unit and deleted at will too
So every time new watcher is added, all store "up" from it must have actual state - but if these stores were not observed before,
their state is statle and must be recalculated. It is easy to find "roots" from which recalculation must be launched, but currently launch of the nodes will also produce effect calls - which is not needed. So, effector kernel must have "dry run" feature - only state is recalculated, effects are not called

Also there is allSettled - in ssr we calculate state before we actually know, which parts of state are going to be watched by view-layer, so kernel also must have a switch between "lazy" (unobserved nodes are skipped) and "hot" modes of calculation
NOTE: it would be cool, if watchers were not dynamic too, for e.g. there is no store.watch and getState apis, but separate kind of nodes to represent watchers - this way never-to-be-observed parts of app can be skipped in allSettled and shown in dev-tools